clean up previous commits a bit + refactor most modals into one dir
